requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
numpy==1.26.2
python-dotenv==1.0.0
python-multipart==0.0.6
pyjwt==2.8.0
email-validator==2.1.0

# OpenAI and related
openai>=1.24.0,<2.0.0
tiktoken>=0.7,<1

# LangChain - compatible versions
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.7
langchain-google-genai==0.0.8
langchain-text-splitters==0.0.1

# Google AI
google-generativeai==0.3.2

# Database
qdrant-client==1.7.0
psycopg2-binary==2.9.9

# Groq
groq==0.4.1

# sentence-transformers dependencies
torch>=2.0.0
transformers>=4.30.0
sentence-transformers==2.2.2

fastembed>=0.2.0